requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pillow==11.3.0
requests==2.32.4
pydantic-settings==2.10.1
torch==2.9.1
torchvision==0.24.1
